JavaEE added as contender
Google Protobuf used to generate interface DTOs. Will be used in the other languages as well.

At the moment JPA is used to access the database but this will be replaced because JPA is to heavyweight and since a pure interface service cannot make use of lazy loading it doesnt make sense.