Simple orchestration layer with Akka HTTP
The initial solution is not reactive, it just creates a approval in the wife-service when a new bike was created in the bike-service.

The plan is to make this layer reactive by using persistent actors which stored failed events/messages for later retry. Take actors out of service if a microservice vanishes etc.