Using Akka-Persistence for command-sourcing incoming events.
All events from external services are persisted in the event-store. If the event was successfully dispatched to the receiver, a successfull-delivery-event will be persisted.
Thus, the actor is able to recreate it's state and knows which messages need to be delivered (retried).

Serialization of the commands to the LevelDB has been configured by using Kyro instead of Java-Serialization